# src/main/resources/static/api.yaml
openapi: 3.0.3

info:
  title: pHelix Delivery Manager
  version: v0.1

servers:
  - url: http://localhost:8080/api/v1

paths:

  /delivery-jobs:
    get:
      tags:
        - Jobs
      description: List delivery jobs
      operationId: listJobs
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/sortParam'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeliveryJob'

    post:
      tags:
        - Jobs
      description: Create a new Transfer Job
      operationId: createJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeliveryJob'
      parameters:
        - in: query
          name: allowDuplication
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeliveryJob'

  /delivery-jobs/{id}:
    get:
      tags:
        - Jobs
      description: Get single job
      operationId: getJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeliveryJob'

  /connections:
    get:
      tags:
        - Connections
      description: List connections
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/sortParam'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Connection'

    post:
      tags:
        - Connection
      description: Create a new connection
      operationId: createConnection
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConnection'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Connection'

  /connections/{id}:
    put:
      tags:
        - Connection
      description: Update connection
      operationId: updateConnection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConnection'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Connection'

  /profiles:
    get:
      tags:
        - Profiles
      description: List profiles
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/sortParam'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Profile'

    post:
      tags:
        - Profiles
      description: Create a new profile
      operationId: createProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfile'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'

  /profiles/{id}:
    put:
      tags:
        - Profiles
      description: Update profile
      operationId: updateProfile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfile'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'

  /contacts:
    get:
      tags:
        - Contacts
      description: List contacts
      operationId: listContacts
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/sortParam'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'

    post:
      tags:
        - Contacts
      description: Create a new contact
      operationId: createContact
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContact'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'

  /contacts/{id}:
    put:
      tags:
        - Contacts
      description: Update contact
      operationId: updateContact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContact'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'

  /storages:
    get:
      tags:
        - Storages
      description: List storages
      operationId: listStorages
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/sortParam'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Storage'

    post:
      tags:
        - Storages
      description: Create a new storage
      operationId: createStorage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStorage'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Storage'

  /storages/{id}:
    put:
      tags:
        - Storages
      description: Update storage
      operationId: updateStorage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStorage'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Storage'

  /email-notifications:
    post:
      tags:
        - Email notifications
      description: Create a new email notification
      operationId: createEmailNotification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailNotification'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailNotification'

components:
  parameters:
    offsetParam:
      in: query
      name: offset
      required: false
      schema:
        type: integer
        minimum: 0
      description: The number of items to skip before starting to collect the result set.
    limitParam:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: The numbers of items to return.
    filterParam:
      in: query
      name: filter
      required: false
      schema:
        type: string
      description: RSQL Filter
    sortParam:
      in: query
      name: sort
      required: false
      schema:
        type: string
        default: 'id,desc'
      description: RSQL Sort

  schemas:
    Error:
      type: object
      description: General Error
      properties:
        message:
          type: string
          description: Error message
        constraintViolations:
          type: array
          items:
            $ref: '#/components/schemas/ConstraintViolation'
        errorMessages:
          type: array
          description: Simple error message
          items:
            type: string
      required:
        - message

    ConstraintViolation:
      type: object
      description: Constraint violation
      properties:
        message:
          type: string
          description: Error message
        property:
          type: string
          description: Property path

#   Asset
    Asset:
      type: object
      description: Transfer Asset
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: ID.
        path:
          type: string
          description: Path of the asset
        phelixAssetId:
          type: integer
          format: int64
          description: The pHelix ID of the asset.
        phelixAssetVersion:
          type: integer
          format: int64
          description: The pHelix Version ID of the asset.

#   Delivery job
    DeliveryJob:
      type: object
      description: Transfer Job
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: ID.
        profile:
          $ref: '#/components/schemas/Profile'
        assets:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        titleId:
          type: string
          description: pHelix Title ID
        alphaId:
          type: integer
          format: int64
          description: pHelix Alpha ID
        status:
          type: string
          enum:
            - CREATED
            - FAILED
            - CANCELED
            - IN_PROGRESS
            - COMPLETED
        progress:
          type: number
          format: double
        initiator:
          type: string
          enum:
            - OM
        reportingId:
          type: string
        emailSubject:
          type: string
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/DeliveryJobContact'
        name:
          type: string
        externalSystemId:
          type: string
        taskOffice:
          type: string
        omJobOffice:
          type: string
        owner:
          type: string
        createdAt:
          type: string
        startedAt:
          type: string
        completedAt:
          type: string
        autoNotify:
          type: boolean
          default: true
      required:
        - assets
        - profile

    CreateDeliveryJob:
      type: object
      description: Transfer Job
      properties:
        profile:
          $ref: '#/components/schemas/BasicEntity'
        assets:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        initiator:
          type: string
          enum:
            - OM
        reportingId:
          type: string
        emailSubject:
          type: string
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/CreateDeliveryJobContact'
        name:
          type: string
        externalSystemId:
          type: string
        taskOffice:
          type: string
        omJobOffice:
          type: string
        autoNotify:
          type: boolean
          default: true
      required:
        - assets
        - profile

#   Connection
    Connection:
      type: object
      description: Connection
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: ID.
        name:
          type: string
          description: Connection name
        description:
          type: string
          description: Connection description
        type:
          type: string
          description: Connection type (e.g. Aspera...)
          enum:
            - ASPERA
        config:
          type: object
          description: Connection configurations in JSON format
      required:
        - name
        - type
        - config

    CreateConnection:
      type: object
      description: CreateConnection
      properties:
        name:
          type: string
          description: Connection name
        description:
          type: string
          description: Connection description
        type:
          type: string
          description: Connection type (e.g. Aspera...)
        config:
          type: object
          description: Connection configurations in JSON format
      required:
        - name
        - type
        - config

#   Contact
    Contact:
      type: object
      description: Contact
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: ID.
        firstName:
          type: string
          description: Contact first name
        lastName:
          type: string
          description: Contact last name
        email:
          type: string
          format: email
          description: Contact first name
      required:
        - email

    CreateContact:
      type: object
      description: CreateContact
      properties:
        firstName:
          type: string
          description: Contact first name
        lastName:
          type: string
          description: Contact last name
        email:
          type: string
          format: email
          description: Contact first name
      required:
        - email

#   Profile
    ProfileContact:
      type: object
      description: ProfileContact
      properties:
        contact:
          $ref: '#/components/schemas/Contact'
#        profile:
#          $ref: '#/components/schemas/Profile'
        notifyOn:
          type: string
          enum:
            - SUCCESS
            - FAILURE
          description: notify on (SUCCESS or FAILURE)
        sendType:
          type: string
          enum:
            - TO
            - CC
            - BCC
          description: TO,CC or BCC
      required:
        - notifyOn
        - sendType

    CreateProfileContact:
      type: object
      description: CreateProfileContact
      properties:
        contact:
          $ref: '#/components/schemas/BasicEntity'
        notifyOn:
          type: string
          enum:
            - SUCCESS
            - FAILURE
          description: notify on (SUCCESS or FAILURE)
        sendType:
          type: string
          enum:
            - TO
            - CC
            - BCC
          description: TO,CC or BCC
      required:
        - notifyOn
        - sendType

#   Delivery job contact
    DeliveryJobContact:
      type: object
      description: DeliveryJobContact
      properties:
        contact:
          $ref: '#/components/schemas/Contact'
#        deliveryJob:
#          $ref: '#/components/schemas/DeliveryJob'
        notifyOn:
          type: string
          enum:
            - SUCCESS
            - FAILURE
          description: notify on (SUCCESS or FAILURE)
        sendType:
          type: string
          enum:
            - TO
            - CC
            - BCC
          description: TO,CC or BCC
      required:
        - notifyOn
        - sendType

    CreateDeliveryJobContact:
      type: object
      description: CreateDeliveryJobContact
      properties:
        contact:
          $ref: '#/components/schemas/BasicEntity'
        notifyOn:
          type: string
          enum:
            - SUCCESS
            - FAILURE
          description: notify on (SUCCESS or FAILURE)
        sendType:
          type: string
          enum:
            - TO
            - CC
            - BCC
          description: TO,CC or BCC
      required:
        - contact
        - notifyOn
        - sendType

#   Profile
    Profile:
      type: object
      description: Connection Profile
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: ID.
        name:
          type: string
          description: Connection name
        description:
          type: string
          description: Connection description
        connection:
          $ref: '#/components/schemas/Connection'
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/ProfileContact'
        rootPath:
          type: string
          description: Root path in the destination
          default: "/"
        emailSubjectTemplate:
          type: string
          description: template that will be resolved upon job creation
        completionTemplateKey:
          type: string
        failureTemplateKey:
          type: string
      required:
        - name
        - connection

    CreateProfile:
      type: object
      description: CreateProfile
      properties:
        name:
          type: string
          description: Connection name
        description:
          type: string
          description: Connection description
        connection:
          $ref: '#/components/schemas/BasicEntity'
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/CreateProfileContact'
        rootPath:
          type: string
          description: Root path in the destination
          default: "/"
        emailSubjectTemplate:
          type: string
          description: template that will be resolved upon job creation
        completionTemplateKey:
          type: string
        failureTemplateKey:
          type: string
      required:
        - name
        - connection

#   Storage
    Storage:
      type: object
      description: Storage Mapping
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: ID.
        name:
          type: string
          description: Storage name
        storageId:
          type: string
          description: Storage AM ID
        location:
          type: string
          description: Storage location
        attributes:
          $ref: '#/components/schemas/StorageAttributes'
      required:
        - storageId

    CreateStorage:
      type: object
      description: Create Storage
      properties:
        name:
          type: string
          description: Storage name
        storageId:
          type: string
          description: Storage AM ID
        location:
          type: string
          description: Storage location
        attributes:
          $ref: '#/components/schemas/StorageAttributes'
      required:
        - storageId
        - attributes

    StorageAttributes:
      type: object
      additionalProperties: true

#   Email
    EmailNotification:
      type: object
      description: email notification mapping
      properties:
        type:
          type: string
          enum:
            - Completion
            - Failure
        initiator:
          type: string
          enum:
            - OM
        externalId:
          type: string
        deliveryJob:
          type: object
          properties:
            id:
              type: integer
              format: int64
      required:
        - deliveryJob
        - type

#   Common
    BasicEntity:
      type: object
      properties:
        id:
          type: integer
          format: int64
      required:
        - id